Extract field selection application logic into a hook
